rgw/d4n: squashing all commits related to initial implementation of write-back cache for non-multipart objects (writing objects to cache and then to backend store in the cleaning process):
1. combining write cache with latest D4N code
2. cleaning the code
3. removing some bugs on bigger objects
4. updating iterate function to check the dirty flag
5. Updating write cache based on the pull request comments.
Read process needs to be updated based on write process. It needs to check where is the data and if it is dirty or clean.
If it is in the cache and dirty, we need to put D_ in the oid of the object before reading it from cache.
If it is clean, there is nothing to do.
6. updating flush functions and comments.

Signed-off-by: mosayyebzadeh <mosayyeb@bu.edu>